Split up writing of data summarisation and reading of retrainable model
